Remove resetting registry after reports have been exported
Due to the resetting the registry after reports have been exported
http_server has only the results of the first export. So instead of
resetting the registry we check whether the given Gauge can be found it
it and if yes, we simply return if from `_names_to_collectors`
dictionary; otherwise we build it from scratch.

* Resetting registry is removed from `main`
* `gaarf-exporter` is bumped to 0.3.0 version

Change-Id: I6eb045513b74265d01e64585474854782628a65a